app/backend/pyproject.toml
[project]
name = "portfolio-api"
version = "0.1.0"
description = "Backend API for portfolio website."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.34,<2.0",
  "email-validator>=2.1,<3.0",
  "fastapi>=0.111,<1.0",
  "punq>=0.7,<1.0",
  "pydantic-settings>=2.2,<3.0",
  "ulid-py>=1.1,<2.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
